requirements.txt
langchain
python-dotenv
langchain_community
langchain_core
streamlit
bs4
arxiv
pymupdf
wikipedia
chromadb
langchain-text-splitters
ipykernel
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
duckdb
pandas
openai
langchain_groq
pypdf